UPDATE: Front-Showcase,Login,FAQs Back-hash.js,index.js

+++ frontend_react/src/pages/Login.js
@@ -2,40 +2,77 @@ import { useEffect, useState } from "react";
 import { useNavigate } from 'react-router-dom';
 import { fetchFromServer } from "../utils/fetchFromServer";
 
-import { Box, Button, Stack, Typography } from "@mui/material"
-import { BG_Login, ErrorIcon, logo, particlesBg, scene, sunflowerBg } from "../utils/constants";
-import Input from "../components/Input";
+import { Box, Button, IconButton, InputAdornment, Link, Stack, TextField, Typography } from "@mui/material"
+import { BG_Login, ErrorIcon, GoogleAuth, VisibilityIcon, VisibilityOffIcon, logo, particlesBg, scene, sub_logo, sunflowerBg } from "../utils/constants";
 import { ErrorMessage } from "../components";
 
 
 const Login = () => {
-  const initialState = { email_username: '', password: '' };
+  const initialState = { email_username: '', password: '', confirmPassword: '' };
   const [formData, setFormData] = useState(initialState);
-  const [showPassword, setShowPassword] = useState(false);
+  const [typeInput, setTypeInput] = useState("password");
   const [message, setMessage] = useState('');
+  const [isSignUp, setIsSignUp] = useState(false);
+  const [isResetPassword, setIsResetPassword] = useState(false);
   const navigate = useNavigate();
 
   const handleSubmit = async (e) => {
     e.preventDefault();
 
     // console.log(formData);
-    fetchFromServer('login', formData)
-    .then((res) => {
-      if(res.data) {
-        localStorage.setItem('user', JSON.stringify(res.data));
-        navigate('/home');
-      } else {
-        setMessage('Email ou mot de passe incorrects')
-      }
-    })
-    .catch((err) => {
-      setMessage('Email ou mot de passe incorrects');
-    });
+    if(!formData.confirmPassword) {
+      fetchFromServer('login', formData)
+      .then((res) => {
+        if(res.data) {
+          localStorage.setItem('user', JSON.stringify(res.data));
+          navigate('/home');
+        } else {
+          setMessage('Email ou mot de passe incorrects')
+        }
+      })
+      .catch((err) => {
+        setMessage('Email ou mot de passe incorrects');
+      });
+    } else {
+      fetchFromServer('register', formData)
+      .then((res) => {
+        if(res.data) {
+          localStorage.setItem('user', JSON.stringify(res.data));
+          navigate('/home');
+        } else {
+          setMessage('Email ou mot de passe incorrects')
+        }
+      })
+      .catch((err) => {
+        setMessage('Email ou mot de passe incorrects');
+      });
+    }
   }
 
   const handleChange = (e) => {
     setFormData({ ...formData, [e.target.name]: e.target.value });
   }
+
+  const switchToSignUp = (e) => {
+    setIsSignUp(isSignUp => !isSignUp);
+  }
+
+  const switchToResetPassword = (e) => {
+    setIsResetPassword(isResetPassword => !isResetPassword);
+  }
+
+  const handleShowPassword = () => {
+    if(typeInput === "text"){
+      setTypeInput("password");
+    } else {
+      setTypeInput("text");
+    }
+  }
+
+  // const switchMode = () => {
+  //   setIsSignUp((signedUp) => !signedUp);
+  //   setShowPassword(false);
+  // }
 
   useEffect(() => {
     const user = JSON.parse(localStorage.getItem('user'));
@@ -48,50 +85,103 @@ const Login = () => {
 
   }, [navigate]);
 
-  const handleShowPassword = () => setShowPassword((seePassword) => !seePassword);
-
-  // const switchMode = () => {
-  //   setIsSignUp((signedUp) => !signedUp);
-  //   setShowPassword(false);
-  // }
-
 
 
   return (
-    <Stack sx={{ top: 0, right: 0, display: "flex", justifyContent: "center", alignItems: "center", width: "100%", height: "100dvh", backgroundImage: `url('${BG_Login}')`, backgroundSize: "cover", backgroundPosition: "center" }}>
-      <Box sx={{ display: "flex" }}>
-        <Box sx={{ display: "flex", justifyContent: "center", alignItems: "center", backgroundImage: `url('${sunflowerBg}')`, backgroundSize: "cover", backgroundPosition: "center", width: "30rem", borderRadius: "20px 0 0 20px", padding: "60px" }}>
-          <Stack justifyContent="center" alignItems="center" component="form" padding="20px" onSubmit={handleSubmit} spacing={8} width="100%">
-            <Box sx={{ display: "flex", flexDirection: "column", justifyContent: "center", alignItems: "center", gap: 5, width: "80%" }}>
-              <Typography variant="h1" sx={{ fontSize: "2.5em", fontWeight: "600", color: "#152C5B" }}>Connexion</Typography>
-              <Stack spacing="20px" width="100%">
-                {message !== '' && <ErrorMessage message={message} />}
-                {/* {isSignUp && (
-                  <>
-                    {message !== '' && (
-                      <div style={{ backgroundColor: "red", padding: "5px", borderRadius: "10px", color: "#FFF" }}>
-                        <Typography>{message}</Typography>
-                      </div>
-                    )}
-                    <Input name="lastname" label="Nom(s)" handleChange={handleChange} type="text" />
-                    <Input name="firstname" label="Prénom(s)" handleChange={handleChange} type="text" />
-                  </>
-                )} */}
-                <Input name="email_username" label="Email ou nom d'utilisateur" handleChange={handleChange} type="email" />
-                <Input name="password" label="Mot de passe" handleChange={handleChange} type={showPassword ? "text" : "password"} handleShowPassword={handleShowPassword} />
-                {/* {isSignUp && <Input name="confirmPassword" label="Confirmer le mot de passe" handleChange={handleChange} type="password" />} */}
-              </Stack>
-            </Box>
-            <Button type="submit" variant="contained" size="large" sx={{ width: "50%" }} onSubmit={handleSubmit}>S'identifier</Button>
-            {/* <Button onClick={switchMode}>{isSignUp ? 'Déjà utilisateur ?' : 'Nouvel Utilisateur ?'}</Button> */}
+    <Box sx={{ display: "grid", flexDirection: "column", placeContent: "center", height: "100dvh", width: "100%", backgroundImage: `url('${BG_Login}')`, backgroundSize: "cover", backgroundPosition: "center" }}>
+      {!isResetPassword ?
+        <Stack component="form" sx={{ maxWidth: "24rem", gap: "1.75rem", borderRadius: "5px", padding: "2.5rem", textAlign: "center", /*backgroundImage: `url('${sunflowerBg}')`, backgroundSize: "cover", backgroundPosition: "center"*/backgroundColor: "#FFF" }} onSubmit={handleSubmit}>
+          <Stack alignItems="center">
+            <img src={logo} alt="HKSCRAP_logo" style={{ borderStyle: "none", display: "block", verticalAlign: "middle", maxWidth: "100%" }} height={80} />
           </Stack>
-        </Box>
-
-        <Box sx={{ display: "flex", flexDirection: "column", justifyContent: "center", alignItems: "center", padding: "20px", backgroundImage: `url('${scene}')`, backgroundSize: "cover", backgroundPosition: "center", width: "20rem", borderRadius: "0 20px 20px 0" }}>
-          <img src={logo} alt="scene-haikei" height="40%" />
-        </Box>
-      </Box>
-    </Stack>
+          {!isSignUp ?
+            <Stack spacing="0.5rem">
+              <Typography variant="h1" sx={{ fontFamily: "Montserrat", color: "#37446f", fontSize: "1.25rem", lineHeight: "1.75rem", fontWeight: 700 }}>Connectez-vous à votre compte</Typography>
+              <Typography sx={{ fontFamily: "Montserrat", fontSize: "0.875rem", fontWeight: 600, color: "#969cb6" }}>Bienvenue!<br />Veuillez saisir vos identifiants</Typography>
+            </Stack>
+            :
+            <Stack spacing="0.5rem">
+              <Typography variant="h1" sx={{ fontFamily: "Montserrat", color: "#37446f", fontSize: "1.25rem", lineHeight: "1.75rem", fontWeight: 700 }}>Rejoignez la communauté</Typography>
+              <Typography sx={{ fontFamily: "Montserrat", fontSize: "0.875rem", fontWeight: 600, color: "#969cb6" }}>Inscrivez-vous dès maintenant pour accéder aux outils et aux ressources dont vous avez besoin pour explorer le monde des données sur le web.</Typography>
+            </Stack>
+          }
+          <Stack spacing="1rem" textAlign="left" minWidth="300px">
+            {message !== '' && <ErrorMessage message={message} />}
+            <TextField required type="email" size="small" placeholder="Email" name="email_username" sx={{ display: "flex", width: "100%", fontFamily: "Montserrat", height: "2.5rem", borderRadius: "0.375rem", fontSize: ".875rem", lineHeight: "1.25rem" }} onChange={handleChange} />
+            <TextField required type={typeInput} size="small" placeholder="Mot de passe" name="password" sx={{ display: "flex", width: "100%", fontFamily: "Montserrat", height: "2.5rem", borderRadius: "0.375rem", fontSize: ".875rem", lineHeight: "1.25rem" }} onChange={handleChange}
+              InputProps={!isSignUp ? {
+                endAdornment: (
+                  <InputAdornment position="end">
+                    <IconButton onClick={handleShowPassword}>
+                      {typeInput === "password" ? <VisibilityOffIcon /> : <VisibilityIcon />}
+                    </IconButton>
+                  </InputAdornment>
+                )
+              }: null}
+            />
+            {isSignUp &&
+              <TextField required type={typeInput} size="small" placeholder="Confirmez le mot de passe" name="confirm_password" sx={{ display: "flex", width: "100%", fontFamily: "Montserrat", height: "2.5rem", borderRadius: "0.375rem", fontSize: ".875rem", lineHeight: "1.25rem" }} onChange={handleChange} />
+            }
+            {!isSignUp ?
+              <Box component={Button} type="submit" sx={{ display: "flex", justifyContent: "center", alignItems: "center", border: "1px solid #e6e6e6", padding: "10px 15px", cursor: "pointer", textDecoration: "none", borderRadius: "10px", backgroundColor: "#eaf0fa", transition: ".2s" }}  onSubmit={handleSubmit}>
+                <Typography sx={{ fontFamily: "Montserrat", fontSize: "16px", fontWeight: 700, color: "#0056b3", textTransform: "uppercase" }}>Connexion</Typography>
+              </Box>
+              :
+              <Box component={Button} type="submit" sx={{ display: "flex", justifyContent: "center", alignItems: "center", border: "1px solid #e6e6e6", padding: "10px 15px", cursor: "pointer", textDecoration: "none", borderRadius: "10px", backgroundColor: "#0056b3", transition: ".2s" }}  onSubmit={handleSubmit}>
+                <Typography sx={{ fontFamily: "Montserrat", fontSize: "16px", fontWeight: 700, color: "#FFF", textTransform: "uppercase" }}>Je m'inscris</Typography>
+              </Box>
+            }
+          </Stack>
+          <Stack>
+            <Box sx={{ marginBottom: "0.75rem", position:"relative" }}>
+              <hr style={{ position: "absolute", left: 0, right: 0, top: "17%", color: "#f2f2f2", boxSizing: "content-box", overflow: "visible", height: 0, borderTopWidth: "1px" }} />
+              <Typography sx={{ display: "inline-block", position: "relative", backgroundColor: "#FFF", zIndex: 10, padding: "0 .5rem", fontSize: "12px", fontFamily: "Montserrat", color: "#969cb6", lineHeight: "1rem", fontWeight: 600 }}>{!isSignUp ? 'ou se connecter avec' : 'ou s\'inscrire avec'}</Typography>
+            </Box>
+            <Box sx={{ display: "inline-flex", alignSelf: "center", justifyContent: "center", width: "20px", height: "20px", backgroundColor: "#FFF", border: "1px solid #e1e1e1", borderRadius: "5px", padding: "10px"  }}>
+              <img src={GoogleAuth} alt="GoogleOAuth" style={{ display: "block", width: "100%", height: "100%" }} />
+            </Box>
+          </Stack>
+          <Stack justifyContent="center" spacing="0.25rem" sx={{ minHeight: "40px" }}>
+            {!isSignUp ?
+              <>
+              <Box sx={{ display: "flex", justifyContent: "center" }}>
+                <Typography fontSize=".75rem" lineHeight="1rem" fontFamily="Montserrat" color="#8c8c8c">Pas de compte?</Typography>
+                <Typography component={Link} sx={{ fontSize: ".75rem", lineHeight: "1rem", fontFamily: "Montserrat", fontWeight: 600, color: "#000", textDecoration: "none", cursor: "pointer" }} onClick={switchToSignUp}>S'inscrire</Typography>
+              </Box>
+              <Box sx={{ display: "flex", justifyContent: "center" }}>
+                <Typography fontSize=".75rem" lineHeight="1rem" fontFamily="Montserrat" color="#8c8c8c">Mot de passe oublié?</Typography>
+                <Typography component={Link} sx={{ fontSize: ".75rem", lineHeight: "1rem", fontFamily: "Montserrat", fontWeight: 600, color: "#000", textDecoration: "none", cursor: "pointer" }} onClick={switchToResetPassword}>Cliquez ici</Typography>
+              </Box>
+              </>
+              :
+              <Box sx={{ display: "flex", justifyContent: "center" }}>
+                <Typography fontSize=".75rem" lineHeight="1rem" fontFamily="Montserrat" color="#8c8c8c">Déjà utilisateur?</Typography>
+                <Typography component={Link} sx={{ fontSize: ".75rem", lineHeight: "1rem", fontFamily: "Montserrat", fontWeight: 600, color: "#000", textDecoration: "none", cursor: "pointer" }} onClick={switchToSignUp}>Se connecter</Typography>
+              </Box>
+            }
+          </Stack>
+        </Stack>
+        :
+        <Stack component="form" sx={{ maxWidth: "24rem", gap: "1.75rem", borderRadius: "5px", padding: "2.5rem", textAlign: "center", /*backgroundImage: `url('${sunflowerBg}')`, backgroundSize: "cover", backgroundPosition: "center"*/backgroundColor: "#FFF" }} onSubmit={handleSubmit}>
+          <Stack spacing="0.5rem">
+            <Typography variant="h1" sx={{ fontFamily: "Montserrat", color: "#37446f", fontSize: "1.25rem", lineHeight: "1.75rem", fontWeight: 700 }}>Réinitialisez votre mot de passe</Typography>
+            <Typography sx={{ fontFamily: "Montserrat", fontSize: "0.875rem", fontWeight: 600, color: "#969cb6" }}>Vous recevrez un lien à l'adresse mail</Typography>
+          </Stack>
+          <Stack spacing="1rem" textAlign="left" minWidth="300px">
+            <TextField required type="email" size="small" placeholder="Email" name="email_username" sx={{ display: "flex", width: "100%", fontFamily: "Montserrat", height: "2.5rem", borderRadius: "0.375rem", fontSize: ".875rem", lineHeight: "1.25rem" }} onChange={handleChange} />
+            <Box component={Button} type="submit" sx={{ display: "flex", justifyContent: "center", alignItems: "center", border: "1px solid #e6e6e6", padding: "10px 15px", cursor: "pointer", textDecoration: "none", borderRadius: "10px", backgroundColor: "#0056b3", transition: ".2s" }}  onSubmit={handleSubmit}>
+                <Typography sx={{ fontFamily: "Montserrat", fontSize: "16px", fontWeight: 700, color: "#FFF", textTransform: "uppercase" }}>Envoyer le lien</Typography>
+              </Box>
+          </Stack>
+          <Stack justifyContent="center" spacing="0.25rem" sx={{ minHeight: "40px" }}>
+            <Box sx={{ display: "flex", justifyContent: "center" }}>
+              <Typography fontSize=".75rem" lineHeight="1rem" fontFamily="Montserrat" color="#8c8c8c">Identifiants retrouvés?</Typography>
+              <Typography component={Link} sx={{ fontSize: ".75rem", lineHeight: "1rem", fontFamily: "Montserrat", fontWeight: 600, color: "#000", textDecoration: "none", cursor: "pointer" }} onClick={switchToResetPassword}>Se connecter</Typography>
+            </Box>
+          </Stack>
+        </Stack>
+      }
+    </Box>
   )
 }
